Add nextExclusiveSignedThing() to EnhancedRandom.
Where Thing is Float or Double. These are efficient ways to generate numbers that can be close to 0 but not be 0, and can be positive or negative. Many easy solutions to this task force the lowest mantissa bit or bits to be 0, while this keeps all mantissa bits equally likely to be set.