Fix some things in DistributedRandom.
Its nextExclusiveDouble() method called nextLong(long), which isn't currently defined in a way that uses the distribution, and I'm not sure how I even would define it properly. Now it uses a distributed double much more directly. There are also exclusive signed versions of the float and double code, which just set the sign based on the parity of the generator's state (multiplied by a large constant).