Benchmark JDK19's Math.unsignedMultiplyHigh().
Wow, it's quite a lot faster with its intrinsics than we currently get in nextLong(), and the math is the same (it even had distribution issues on the same limit as nextLong(), though it's only a tiny bit under my estimate of where it should have been).

This makes distributor currently need JDK19 (or higher) to build, but the main library still uses JDK8.