I think this is a bug in previousLong().
There was no way to retrieve the long that had just been returned by nextLong(), which is pretty critical to doing something like undoing a shuffle. I've already fixed it in TupleQuasiRandom, since the fix there is extremely simple, but I'll want to go through and change every previousLong() implementation.