Don't need MX3 if I've got an XLCG!
This corrects a notable omission in Ziggurat, which was probably the cause for the strange "fat lip" at the start of the trail section in some earlier versions -- x needs to be divided by R. Once that was added, I noticed the range was smaller, which makes sense. By fiddling with quick randomization schemes for a few hours, I found using the high bits of an XLCG seems to work well.