Adjustments to PouchRandom.
Some of these help the code stay robust in case of problem input, and some of the changes are temporary attempts to reduce correlation between similar starting states early on. The temporary attempts will be removed, since they make that version instantly fail PractRand, but it's good to at least know one way that reduces correlation.