Try to reduce hex float usage in copy-able code.
Kotlin, Scala, and Clojure don't have hex float literals (not to mention C# and JS), so copying code that uses them is tricky. I'm trying to walk a fine line where I have the hex versions available in a comment, since that makes it easier for me to edit them, but I use a decimal version in the code.