HashMap works, but jdkgdxds sorta doesn't.
The jdkgdxds IntObjectMap throws an Exception when multiple iterators try to go over the map at the same time, but not if one or more was created by the Entries constructor. Removal via the Entries iterator doesn't notify the rest of the iterators that the map has changed, and so removing using those won't work unless they call next(). HashMap has this work fine; I'm not sure exactly what it does differently.