Very interesting ICE test results...
PcgRXSMXSRandom is expected to fail when given sequential y values as its stateB, since it only accepts odd numbers for stateB. But what surprised me was that if stateA is initialized only with even numbers, there is a very slightly detectable correlation, enough to make it fail this test. Initializing stateA with odd numbers seems usually fine.