Add Chock32Random as a simpler Choo32.
So far, it's doing well in PractRand (though it's only just started, 256GB without anomalies is a nice way to begin), and it passes its first try at the ICE test (not the immediate test, probably, but I haven't run it yet). There are also some doc corrections in Choo32Random. Fury has been updated to 0.10.0 .