Avoid output being dependent on new states.
(Applies to Chock32Random.) This is a "trick" used by various generators to improve ILP. I don't know how much it will help here, but it should be an optimization for nextInt() and nextLong() but slightly slower for previousInt() and previousLong(), which is probably the right thing to do. This also fixes the tag for Chock32Random, and implements nextUnsignedInt() using nextInt().