Update digital and... uh...
There is a new-ish Base constant, BASE90, that does include backtick and tilde as digits. This wreaks some havoc with our existing string serialization, and will need some fixing for special cases... We can use the base.paddingChar and base.positiveSign, both of which are never used in a normal generator's serialized form, as the substitutes for backtick and tilde, respectively, but it will be a backwards-incompatible change and will need to be changed in several places.